extract ModalViewHolder from Modal model
